EnemyAttack usage test done | Weapon.cs updated (+creator +config)

+++ Assets/_Scripts/ITEM/WeaponStats.cs
@@ -19,8 +19,8 @@ public class WeaponStats {
 	public int elementDamageBonus; // added to other hand attacks if same element
 	public Element element;
 
-	public int dodgeModifier; // may be negative, from -100 to 100?
-	public int blockModifier; // may be negative
+	public float dodgeModifier; // may be negative, from -100 to 100?
+	public float blockModifier; // may be negative
 
 
 }
